+++ pkg/providers/property/testdata/TestDataDomainOwnershipDomains/domains.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_domainownership_domains" "test" {
+}