+++ CHANGELOG.md
@@ -58,6 +58,12 @@
 
 #### FEATURES/ENHANCEMENTS:
 
+
+* DomainOwnership:
+  * Added new data sources:
+    * `akamai_domainownership_domains` - reads details for all the domains.
+    
+
 * mTLS Keystore:
     * Added support for `preferred_ca` in the `akamai_mtlskeystore_client_certificate_akamai` resource.